287 Move sidebar list into separate component
