+++ src/components/sidebar/index.js
@@ -2,11 +2,10 @@ import React from "react";
 import { Link } from "react-router";
 
 // Children
-import SidebarSelectableItem from "./selectable-item";
+import SidebarList from "./list";
 import SidebarSearchInput from "./search-input";
 import { sidebarContent } from "./content";
 import search from "./search";
-import Icon from "../icon";
 
 const searchIndex = search.createSearchIndex(sidebarContent);
 
@@ -28,103 +27,6 @@ class Sidebar extends React.Component {
     });
   }
 
-  renderList(items) {
-    return items
-      .filter((item) => this.isMatchingNode(item))
-      .map((item) => {
-        const toc = item.children.filter((t) => this.isMatchingNode(t));
-        const alwaysExpand = !!this.state.searchTerm && toc.length > 0;
-
-        return (
-          <SidebarSelectableItem
-            key={item.slug}
-            path={`/${item.route}/${item.slug}`}
-            text={item.text}
-            toc={toc}
-            location={this.props.location}
-            alwaysExpand={alwaysExpand}
-          />
-        );
-      });
-  }
-
-  isMatchingNode(node) {
-    return search.isInMatching(node, this.state.matchingNodes);
-  }
-
-  renderContent() {
-    const content = sidebarContent
-      .filter((heading) => heading.render !== false)
-      .filter((heading) => this.isMatchingNode(heading))
-      .map((heading, i) => {
-        const subheadings = heading.children
-          .filter((subheading) => {
-            return subheading.text ? this.isMatchingNode(subheading) : true;
-          })
-          .map((subheading, subheadingIndex) => {
-            return (
-              <div
-                key={subheading.text || `${i}-${subheadingIndex}`}
-              >
-                <p className="Sidebar-SubHeading SubHeading">
-                  {subheading.text}
-                </p>
-                <ul className="Sidebar-List">
-                  {this.renderList(subheading.children)}
-                </ul>
-              </div>
-            );
-          });
-
-        return (
-          <div key={heading.text} className="Sidebar-Grid-block">
-            <p className="Sidebar-Heading">
-              {heading.text}
-            </p>
-            {subheadings}
-          </div>
-        );
-      });
-
-    return (
-      <div className="Sidebar-Grid">
-        { this.isMatchingNode({ text: "Introduction" }) ?
-          <p className="Sidebar-Heading u-noPadding">
-            Introduction
-          </p>
-          : null
-        }
-        <ul className="Sidebar-List">
-          { this.isMatchingNode({ text: "Getting Started" }) ?
-            <li key="sidebarlink-index" className="Sidebar-List-Item">
-              <Link to="/docs" activeClassName="is-active">
-                Getting Started <Icon glyph="internal-link" />
-              </Link>
-            </li>
-            : null
-          }
-          { this.isMatchingNode({ text: "Native" }) ?
-            <li key="sidebarlink-native" className="Sidebar-List-Item">
-              <Link to="/docs/native" activeClassName="is-active">
-                Native <Icon glyph="internal-link" />
-              </Link>
-            </li>
-            : null
-          }
-          { this.isMatchingNode({ text: "Contributing" }) ?
-            <li key="sidebarlink-contributing" className="Sidebar-List-Item">
-              <a href="https://github.com/FormidableLabs/victory/#contributing">
-                Contributing <Icon glyph="external-link" />
-              </a>
-            </li>
-            : null
-          }
-        </ul>
-        {content}
-      </div>
-    );
-  }
-
   render() {
     /* eslint-disable max-len */
     return (
@@ -136,7 +38,12 @@ class Sidebar extends React.Component {
               onChange={this.handleSearch}
             />
           </div>
-          {this.renderContent()}
+          <SidebarList
+            content={sidebarContent}
+            matchingNodes={this.state.matchingNodes}
+            isSearching={!!this.state.searchTerm}
+            location={this.props.location}
+          />
         </nav>
       </div>
     );
